confirmation applied for change page

+++ src/components/allCards/AllCardSection.js
@@ -7,24 +7,19 @@ import Fab from "@mui/material/Fab";
 import AddIcon from "@mui/icons-material/Add";
 import { ThemeProvider } from "@mui/material/styles";
 import { theme } from "../../theme";
-import TextField from "@mui/material/TextField";
 import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogContentText from "@mui/material/DialogContentText";
-import DialogTitle from "@mui/material/DialogTitle";
-import Button from "@mui/material/Button";
 import "../../pages/app/CreateProject.css";
+import CreateDialog from "../createAndEdit/CreateDialog";
+import DialogActions from "@mui/material/DialogActions";
+import Button from "@mui/material/Button";
 
 class AllCardSection extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      title: "",
-      technology: "",
-      deadline: "",
-      description: "",
       createDailog: false,
+      editDailog: false,
+      deleteCheck: false,
     };
   }
   componentDidMount() {
@@ -32,42 +27,35 @@ class AllCardSection extends Component {
   }
 
   handleDelete = (id) => {
+    if (!window.confirm("Are you sure to delete the project ?")) return;
     this.props.deleteProjectAction(id);
-    alert("Click OK to delete");
-    window.location.reload();
+    this.setState({ deleteCheck: true });
+    const timeout = setTimeout(() => {
+      this.setState({ deleteCheck: false });
+    }, 3000);
+
+    return () => {
+      clearTimeout(timeout);
+    };
   };
-  handleClickOpen = () => {
+
+  handleClickOpenCreate = () => {
     this.setState({ createDailog: true });
   };
 
-  handleClose = () => {
+  handleClickOpenEdit = () => {
+    this.setState({ editDailog: true });
+  };
+
+  handleCloseCreate = () => {
     this.setState({ createDailog: false });
   };
 
-  handleChange = (e) => {
-    this.setState({ [e.target.name]: e.target.value });
-  };
-
-  handleSubmit = (e) => {
-    const project = {
-      title: this.state.title,
-      technology: this.state.technology,
-      deadline: this.state.deadline,
-      description: this.state.description,
-    };
-    e.preventDefault();
-    this.props.setprojectAction(project);
-    this.handleClose();
+  handleCloseEdit = () => {
+    this.setState({ editDailog: false });
   };
 
   render() {
-    const { title, technology, deadline, description } = this.state;
-    const enabled =
-      title.length > 0 &&
-      technology.length > 0 &&
-      deadline.length > 0 &&
-      description.length > 0;
-    console.log("this props all card", this.props.projects);
     return (
       <div id="allcard">
         <div id="head">ALL PROJECTS</div>
@@ -76,65 +64,34 @@ class AllCardSection extends Component {
             <Fab
               color="secondary"
               aria-label="add"
-              onClick={this.handleClickOpen}
+              onClick={this.handleClickOpenCreate}
             >
               <AddIcon />
             </Fab>
           </ThemeProvider>
         </div>
         <Dialog open={this.state.createDailog} maxWidth="xl" maxHeight="xl">
-          <div id="headi">+ CREATE PROJECT</div>
-          <div id="formOuter">
-            <form onSubmit={this.handleSubmit} id="formc">
-              <div id="two">
-                <input
-                  type="text"
-                  id="title"
-                  name="title"
-                  placeholder="Project Title"
-                  className="title"
-                  onChange={this.handleChange}
-                  autoComplete="off"
-                  ref={this.emailInputRef}
-                />
-                <input
-                  type="text"
-                  id="techno"
-                  name="technology"
-                  placeholder="Technology "
-                  className="title"
-                  autoComplete="off"
-                  onChange={this.handleChange}
-                />
-              </div>
-              <div id="three">
-                <input
-                  type="text"
-                  id="deadline"
-                  name="deadline"
-                  placeholder="Deadline "
-                  className="title"
-                  autoComplete="off"
-                  onChange={this.handleChange}
-                />
-                <textarea
-                  type="text"
-                  id="description"
-                  name="description"
-                  placeholder="Description "
-                  className="title"
-                  autoComplete="off"
-                  onChange={this.handleChange}
-                />
-              </div>
-              <div id="four">
-                <button disabled={!enabled} type="submit" id="subBut">
-                  Submit
-                </button>
-              </div>
-            </form>
-          </div>
+          <CreateDialog></CreateDialog>
+          <DialogActions>
+            <Button
+              autoFocus
+              onClick={() => {
+                if (
+                  !window.confirm(
+                    "Are you sure to leave, all the changes will be lost?"
+                  )
+                )
+                  return;
+                this.handleCloseCreate();
+              }}
+            >
+              Cancel
+            </Button>
+          </DialogActions>
         </Dialog>
+        {this.state.deleteCheck && (
+          <div id="deleteNotifi">NOTIFICATION : Project is deleted</div>
+        )}
         {this.props.projects.projectData?.reverse().map((res) => {
           return (
             <div id="proDiv">
@@ -169,13 +126,11 @@ AllCardSection.propTypes = {
   projects: PropTypes.object.isRequired,
   getProjectAction: PropTypes.func.isRequired,
   deleteProjectAction: PropTypes.func.isRequired,
-  setprojectAction: PropTypes.func.isRequired,
 };
 
 AllCardSection.defaultProps = {
   getProjectAction: () => {},
   deleteProjectAction: () => {},
-  setprojectAction: () => {},
 };
 
 export default AllCardSection;
